Rename Project > Instances to Project > Overview
